Replace placeholder with lazy loader component.
