+++ src/components/LayoutManager/Components/LazyLoader/LazyLoader.scss
@@ -1,0 +1,7 @@
+.lazyContainer {
+    position: absolute;
+    top: 2px;
+    bottom: 2px;
+    left: 2px;
+    right: 2px;
+}
